New : companyHoliday deletion Change : CompanyHolidayType -> string enum, fixed pipe

+++ src/app/components/company-holiday/company-holiday.component.ts
@@ -1,9 +1,11 @@
 import {AfterViewInit, Component, Inject, OnInit, ViewChild} from '@angular/core';
-import {CompanyHoliday} from "../../models";
+import {CompanyHoliday, CompanyHolidayType, LeaveStatus} from "../../models";
 import {CompanyHolidayService} from "../../services/company-holiday.service";
 import {MatTableDataSource} from "@angular/material/table";
 import {MatSort} from "@angular/material/sort";
-import {MatDialog, MAT_DIALOG_DATA} from '@angular/material/dialog';
+import {MAT_DIALOG_DATA, MatDialog} from '@angular/material/dialog';
+import {MatSnackBar, MatSnackBarConfig} from "@angular/material/snack-bar";
+import {DateTime} from "luxon";
 
 @Component({
   selector: 'app-company-holiday',
@@ -11,6 +13,14 @@ import {MatDialog, MAT_DIALOG_DATA} from '@angular/material/dialog';
   styleUrls: ['./company-holiday.component.scss']
 })
 export class CompanyHolidayComponent implements OnInit, AfterViewInit {
+  LeaveStatus = LeaveStatus;
+  CompanyHolidayType = CompanyHolidayType;
+  snackBarConfig: MatSnackBarConfig = {
+    horizontalPosition: 'center',
+    verticalPosition: 'top',
+    duration: 3000
+  };
+
   years = [ '2022', '2021', '2020', '2019'];
   selectedYear = new Date().getFullYear().toString();
 
@@ -19,7 +29,7 @@ export class CompanyHolidayComponent implements OnInit, AfterViewInit {
 
   @ViewChild(MatSort) sort!: MatSort;
 
-  constructor(private service: CompanyHolidayService, public dialog: MatDialog) { }
+  constructor(private service: CompanyHolidayService, public dialog: MatDialog, private snackBar: MatSnackBar) { }
 
   ngOnInit(): void {
     this.getData();
@@ -48,11 +58,19 @@ export class CompanyHolidayComponent implements OnInit, AfterViewInit {
       data: holiday
     });
 
-    dialogRef.afterClosed().subscribe(result => {
-      console.log(result);
-      this.service.delete(holiday.id)
-        .subscribe(holiday => console.log(holiday));
+    dialogRef.afterClosed().subscribe(confirmed => {
+      if (confirmed)
+        this.service.delete(holiday.id)
+          .subscribe(() => {
+            this.snackBar.open('Suppression effectuée', '', this.snackBarConfig)
+            this.getData();
+          });
     });
+  }
+
+  isDisabled(holiday: CompanyHoliday) {
+    return DateTime.fromJSDate(new Date(holiday.date)) < DateTime.now()
+      || (holiday.type == CompanyHolidayType.COMPANY_RTT && holiday.status == LeaveStatus.VALIDATED);
   }
 }
 
@@ -61,5 +79,7 @@ export class CompanyHolidayComponent implements OnInit, AfterViewInit {
   templateUrl: 'dialog-company-holiday-delete.html'
 })
 export class DialogCompanyHolidayDelete {
+  CompanyHolidayType = CompanyHolidayType;
+
   constructor(@Inject(MAT_DIALOG_DATA) public holiday: CompanyHoliday) {}
 }
